add logging, todo's brewer and grinder consume time dynamically add ExtractionProfiles to encapsulate coffee brew strength add StrengthWanted to Order brewer is told how many ounces to brew

+++ pkg/coffeeshop/beans.go
@@ -1,6 +1,6 @@
 package coffeeshop
 
 type Beans struct {
-	weightGrams int
+	weightGrams float32
 	// indicate some state change? create a new type?
 }
